feat/auth_login: Move LoginForm into Login folder
